feat: rename bytes generator extension
