+++ src/extensions/bytes/src/generator/index.ts
