+++ src/extensions/interfaces/token/src/interface.vue
@@ -85,7 +85,7 @@
 		loading.value = true;
 
 		try {
-			const response = await api.post('/token-generator');
+			const response = await api.post('/generator');
 			emitValue(response.data.data);
 			isNewTokenGenerated.value = true;
 		} catch (err: any) {
